library refactor figurl (step 1)
