+++ src/figurl/index.ts
@@ -1,15 +1,17 @@
 import startListeningToParent from './startListeningToParent'
 
 export { default as getFigureData } from './getFigureData'
+export { default as getFileData, getFileDataUrl, useFileData } from './getFileData'
+export { default as getMutable } from './getMutable'
 export { default as initiateTask, Task } from './initiateTask'
-export { default as runCalculationTaskAsync } from './runCalculationTaskAsync'
 export { default as runActionTaskAsync } from './runActionTaskAsync'
+export { default as runCalculationTaskAsync } from './runCalculationTaskAsync'
 export { default as TaskStatusView } from './TaskStatusView'
 export { default as useCalculationTask } from './useCalculationTask'
-export { default as useSignedIn } from './useSignedIn'
 export { default as useFeed } from './useFeed'
 export { default as useFeedReducer } from './useFeedReducer'
+export { default as useSignedIn } from './useSignedIn'
 export { default as useWindowDimensions } from './useWindowDimensions'
-export { default as validateObject } from './viewInterface/validateObject'
+export { default as validateObject, isArrayOf, isEqualTo, isJSONObject, isNumber, isString, optional } from './viewInterface/validateObject'
 
 startListeningToParent()
